Rename extension to reflect relationship with AsciidoctorJ
The `AsciidoctorExtension` provides the DSL available as `asciidoctorj`
which is used to set the AsciiDoctorJ version. The `asciidoctor` DSL is
provided as methods in the plug-ins. The recommended way to provide DSLs
in Gradle >= 4 is via extensions only. To be able to create a new DSL
extension and not confuse the two extensions, the current one is renamed
to match the fact that only configures AsciiDoctorJ.