Extract source directory setting into extension
Gradle recommends having all configuration in extensions and not in the
task itself. The source directory setting is extracted into the newly
created extension and is passed from the extension to the `asciidoctor`
task.

As a consequence, setting the source directory will only be possible by
using assignments. This is reflected in the README.

This changed needed several adjustments to the tests since the type of
the setting changed. Some were deleted as they are now separate tests in
a specification for the extension. All tests were changed to use
assignments now due to the issue mentioned above.